As what was discussed, I separate the classes, methods and tables into their own java files to be more organized.
